+++ src/org/ironrhino/core/elasticsearch/search/Total.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Total {
 
-	private int value;
+	private long value;
 
 	private String relation;
